Fixed server memory leak. Various small fixes. Various compile fixes.

+++ trunk/MTA10_Server/mods/deathmatch/logic/lua/CLuaFunctionDefinitions.cpp
@@ -441,7 +441,7 @@ int CLuaFunctionDefinitions::AddEventHandler ( lua_State* luaVM )
         uint iPos = strPriority.find_first_of ( "-+" );
         if ( iPos != SString::npos )
         {
-            fPriorityMod = atof ( strPriority.SubStr ( iPos ) );
+            fPriorityMod = static_cast < float > ( atof ( strPriority.SubStr ( iPos ) ) );
             strPriority = strPriority.Left ( iPos );
         }
